+++ packages/moengage_flutter/moengage_flutter_web/CHANGELOG.md
@@ -1,5 +1,11 @@
 # MoEngage Flutter Web Plugin
 
+# Release Date
+
+## Release Version
+
+- [patch] Updated deprecated packages like dart:js
+
 # 03-07-2025
 
 ## 5.0.0